requirements.txt
alembic==1.10.3
anyio==3.6.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.95.1
fastapi-sessions==0.3.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.1
passlib==1.7.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyscopg2==66.0.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
sqlparse==0.4.4
starlette==0.26.1
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.2
zipp==3.15.0
